strace: disable autoreconf
With autoreconf enabled, building strace fails:
Makefile:9303: *** missing separator.  Stop.

Contents of that line:
@CODE_COVERAGE_RULES@

Disabling autoreconf fixes the problem. As the autoreconfig fixup was
enabled in the initial commit of strace to the repo, it's hard to figure
out why it was enabled. It was most likely required to build at that
time, but as it now builds fine without it, just disable the fixup.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>